% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar-iffhs.R
\name{baixar_iffhs}
\alias{baixar_iffhs}
\title{Baixa e limpa o ranking de artilheiros da IFFHS}
\usage{
baixar_iffhs(ano_inicio_carreira = 1900)
}
\arguments{
\item{ano_inicio_carreira}{Ano de início da carreira do jogador}
}
\value{
Um tibble com o ranking de artilheiros da IFFHS
}
\description{
Esta função baixa o ranking de artilheiros da IFFHS e limpa os dados.
}
