% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minha-primeira-funcao.R
\name{calcular_estatisticas}
\alias{calcular_estatisticas}
\title{Calcula estatisticas}
\usage{
calcular_estatisticas(continentes, ano_minimo)
}
\arguments{
\item{continentes}{quais continentes?}

\item{ano_minimo}{qual o ano minimo de filtro?}
}
\value{
base agrupada
}
\description{
Calcula estatisticas
}
