% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{iffhs}
\alias{iffhs}
\title{Data from IFFHS}
\format{
A tibble with 24 rows and 9 columns

\describe{
\item{rank}{Ranking}
\item{player}{Player}
\item{league}{League goals}
\item{cup}{Cup goals}
\item{continental}{Continental goals}
\item{country}{Country goals}
\item{total}{Total goals}
\item{career_span}{Career span}
\item{ano_inicio}{Year of career start}
}
}
\source{
https://en.wikipedia.org/wiki/List_of_footballers_with_500_or_more_goals
}
\usage{
iffhs
}
\description{
Data from IFFHS
}
\keyword{datasets}
