% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minha-primeira-funcao.R
\docType{data}
\name{meus_diamantes}
\alias{meus_diamantes}
\title{meus_diamantes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 53940 rows and 10 columns.
}
\usage{
meus_diamantes
}
\description{
meus_diamantes
}
\keyword{datasets}
